@@ -29,16 +29,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.1.0",
-    "@angular/common": "^6.1.0",
-    "@angular/compiler": "^6.1.0",
-    "@angular/core": "^6.1.0",
-    "@angular/forms": "^6.1.0",
-    "@angular/http": "^6.1.0",
-    "@angular/platform-browser": "^6.1.0",
-    "@angular/platform-browser-dynamic": "^6.1.0",
-    "@angular/platform-server": "^6.1.0",
-    "@angular/router": "^6.1.0",
+    "@angular/animations": "^7.0.4",
+    "@angular/common": "^7.0.4",
+    "@angular/compiler": "^7.0.4",
+    "@angular/core": "^7.0.4",
+    "@angular/forms": "^7.0.4",
+    "@angular/http": "^7.0.4",
+    "@angular/platform-browser": "^7.0.4",
+    "@angular/platform-browser-dynamic": "^7.0.4",
+    "@angular/platform-server": "^7.0.4",
+    "@angular/router": "^7.0.4",
     "ag-grid-angular": "^19.0.0",
     "ag-grid-community": "^19.0.0",
     "bootstrap": "^4.0.0-beta.2",
@@ -47,22 +47,21 @@
     "highcharts-angular": "^2.3.1",
     "highcharts-export-csv": "^1.4.3",
     "ion-rangeslider": "^2.1.7",
-    "ng-recaptcha": "4.0.0-beta.1",
+    "ng-recaptcha": "^4.2.1",
     "popper.js": "^1.12.9",
     "primeng": "^6.0.0-beta.1",
-    "rxjs": "^6.2.2",
-    "rxjs-compat": "^6.2.2",
+    "rxjs": "^6.3.3",
     "tether": "^1.4.0",
     "tlite": "^0.1.6",
     "zone.js": "~0.8.26"
   },
   "devDependencies": {
     "@angular-devkit/architect": "0.7.0-rc.2",
-    "@angular-devkit/build-angular": "0.7.0-rc.3",
+    "@angular-devkit/build-angular": "~0.10.0",
     "@angular-devkit/core": "0.7.0-rc.2",
     "@angular-devkit/schematics": "0.7.0-rc.2",
-    "@angular/cli": "6.2.0-beta.0",
-    "@angular/compiler-cli": "^6.1.0",
+    "@angular/cli": "7.0.6",
+    "@angular/compiler-cli": "^7.0.4",
     "@types/google.analytics": "0.0.33",
     "@types/highcharts": "^5.0.25",
     "@types/jasmine": "~2.8.6",
@@ -82,6 +81,6 @@
     "protractor": "^5.4.0",
     "ts-node": "~7.0.0",
     "tslint": "~5.10.0",
-    "typescript": "2.7.2"
+    "typescript": "3.1.6"
   }
 }
