@@ -1,6 +1,7 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { Component, Input } from '@angular/core';
-import { HttpModule, Http, XHRBackend, BaseRequestOptions, ConnectionBackend, Response, ResponseOptions } from '@angular/http';
+import { XHRBackend, BaseRequestOptions, ConnectionBackend, Response, ResponseOptions } from '@angular/http';
+import { HttpClientModule, HttpClient } from '@angular/common/http';
 import { RouterTestingModule } from '@angular/router/testing';
 import { AnalyzerService } from '../analyzer.service';
 import { HelperService } from '../helper.service';
@@ -22,6 +23,7 @@ class agGridStubComponent {
   @Input() gridReady;
   @Input() gridOptions;
   @Input() suppressDragLeaveHidesColumns;
+  @Input() singleClickEdit;
 }
 
 describe('AnalyzerTableComponent', () => {
@@ -51,7 +53,7 @@ describe('AnalyzerTableComponent', () => {
           }
         }
       ],
-      imports: [ RouterTestingModule, HttpModule ]
+      imports: [ RouterTestingModule, HttpClientModule ]
     })
     .compileComponents();
   }));
