@@ -1,5 +1,7 @@
+import {map} from 'rxjs/operators';
 import { Component, OnInit } from '@angular/core';
 import { Http, Response, Headers, RequestOptionsArgs } from '@angular/http';
+import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { FormGroup, FormBuilder, Validators } from '@angular/forms';
 import { RecaptchaFormsModule } from 'ng-recaptcha/forms';
 
@@ -13,7 +15,7 @@ export class FeedbackComponent implements OnInit {
   public successMsg: string;
   public errorMsg: string;
   private hideAlert: Boolean = true;
-  constructor(private fb: FormBuilder, private http: Http) { }
+  constructor(private fb: FormBuilder, private http: HttpClient) { }
 
   ngOnInit() {
     this.buildForm();
@@ -35,16 +37,14 @@ export class FeedbackComponent implements OnInit {
 
   onSubmit() {
     this.hideAlert = false;
-    const headers = new Headers();
-    headers.append('Authorization', 'Bearer -VI_yuv0UzZNy4av1SM5vQlkfPK_JKnpGfMzuJR7d0M=');
-    const requestOptionsArgs = { headers: headers };
+    const requestOptionsArgs = { headers: new HttpHeaders({ 'Authorization': 'Bearer -VI_yuv0UzZNy4av1SM5vQlkfPK_JKnpGfMzuJR7d0M=' })};
     const feedback = { data: { name: '', email: '', feedback: '' }, 'g-recaptcha-response': '' };
     feedback.data.name = this.feedbackForm.value.name;
     feedback.data.email = this.feedbackForm.value.email;
     feedback.data.feedback = this.feedbackForm.value.feedback;
     feedback['g-recaptcha-response'] = this.feedbackForm.value.captcha;
-    return this.http.post('https://api.uhero.hawaii.edu/v1/feedback', JSON.stringify(feedback), requestOptionsArgs)
-      .map((res: Response) => res.json())
+    return this.http.post('https://api.uhero.hawaii.edu/v1/feedback', JSON.stringify(feedback), requestOptionsArgs).pipe(
+      map((res: Response) => res.json()))
       .subscribe(
       data => this.successMsg = 'Submission successful.',
       error => this.errorMsg = 'Something went wrong. Try again.',
