@@ -123,8 +123,8 @@ export class SingleSeriesComponent implements OnInit, AfterViewInit {
     }
   }
 
-  updateAnalyze(seriesInfo, tableData, chartData) {
-    this._analyzer.updateAnalyzer(seriesInfo.id, tableData, chartData);
+  updateAnalyze(seriesInfo) {
+    this._analyzer.updateAnalyzer(seriesInfo.id);
     seriesInfo.analyze = this._analyzer.analyzerSeries.find(aSeries => aSeries.id === seriesInfo.id);
   }
 
